removing service from routes
