+++ app-connector-client/server/routes/connector.js
@@ -6,7 +6,7 @@ const fs = require("fs")
 const LOGGER = require("../logger").logger
 const forge = require("node-forge")
 const url = require("url")
-const services = require("./services")
+const services = require("../services")
 const events = require("./events")
 const express = require("express")
 const connection = require("../connection")
