+++ app-connector-client/server/routes/mockapis.js
@@ -1,7 +1,7 @@
 const request = require("request")
 const LOGGER = require("../logger").logger
 const express = require("express")
-const services = require("./services")
+const services = require("../services")
 const events = require("./events")
 module.exports = {
     router: router,
@@ -14,21 +14,13 @@ function getAll(req, res) {
     var configApis = varkesConfig.apis;
     for (var i = 0; i < configApis.length; i++) {
         var api = configApis[i]
-        var metadata = services.fillServiceMetadata(api, req.params.hostname);
-        if (api.type == "odata") {
-            metadata.type = "OData";
-        }
-        else {
-            metadata.type = "OpenAPI"
-        }
-        apis.push(metadata)
+
+        apis.push(services.fillServiceMetadata(api, req.params.hostname))
     }
     var configEvents = varkesConfig.events;
     for (var i = 0; i < configEvents.length; i++) {
         var event = configEvents[i];
-        var metadata = events.fillEventData(event);
-        metadata.type = "Event";
-        apis.push(metadata);
+        apis.push(events.fillEventData(event));
     }
     res.status(200).send(apis);
 }
