+++ app-connector-client/server/services.js
@@ -1,17 +1,18 @@
 #!/usr/bin/env node
 'use strict'
 
-const LOGGER = require("../logger").logger
+const LOGGER = require("./logger").logger
 const yaml = require('js-yaml')
 const fs = require("fs")
-const apis = require("./apis")
-
+const connection = require("./connection")
 const OAUTH = "/authorizationserver/oauth/token"
 const METADATA = "/metadata"
 
 module.exports = {
     createServicesFromConfig: createServicesFromConfig,
     getAllAPI: getAllAPI,
+    createAPI: createAPI,
+    updateAPI: updateAPI,
     fillServiceMetadata: fillServiceMetadata
 }
 
@@ -50,7 +51,7 @@ function createService(api, hostname) {
     LOGGER.debug("Auto-register API '%s'", api.name)
     return new Promise((resolve, reject) => {
         var serviceData = fillServiceMetadata(api, hostname)
-        apis.createAPI(serviceData, function (err, httpResponse, body) {
+        createAPI(serviceData, function (err, httpResponse, body) {
             if (err) {
                 reject(err)
             } else {
@@ -70,7 +71,7 @@ function updateService(api, api_id, hostname) {
     LOGGER.debug("Auto-update API '%s'", api.name)
     return new Promise((resolve, reject) => {
         var serviceData = fillServiceMetadata(api, hostname)
-        apis.updateAPI(serviceData, api_id, function (err, httpResponse, body) {
+        updateAPI(serviceData, api_id, function (err, httpResponse, body) {
             if (err) {
                 reject(err)
             } else {
@@ -99,6 +100,39 @@ function getAllAPI() {
                 resolve(JSON.parse(body))
             }
         })
+    })
+}
+function createAPI(serviceMetadata, cb) {
+    request.post({
+        url: connection.info().metadataUrl,
+        headers: {
+            "Content-Type": "application/json"
+        },
+        json: serviceMetadata,
+        agentOptions: {
+            cert: connection.certificate(),
+            key: connection.privateKey()
+        },
+        rejectUnauthorized: connection.secure()
+    }, function (error, httpResponse, body) {
+        cb(error, httpResponse, body)
+    })
+}
+
+function updateAPI(serviceMetadata, api_id, cb) {
+    request.put({
+        url: `${connection.info().metadataUrl}/${api_id}`,
+        headers: {
+            "Content-Type": "application/json"
+        },
+        json: serviceMetadata,
+        agentOptions: {
+            cert: connection.certificate(),
+            key: connection.privateKey()
+        },
+        rejectUnauthorized: connection.secure()
+    }, function (error, httpResponse, body) {
+        cb(error, httpResponse, body)
     })
 }
 
@@ -139,6 +173,7 @@ function fillServiceMetadata(api, hostname) {
 
     if (api.type == "odata") {
         apiData.apiType = "odata"
+
     }
 
     if (!api.type || api.type == "openapi") {
@@ -163,6 +198,7 @@ function fillServiceMetadata(api, hostname) {
         name: api.name,
         description: api.description ? api.description : api.name,
         labels: api.labels ? api.labels : {},
+        type: api.type == "odata" ? "OData" : "OpenAPI",
         api: apiData
     }
